refactor(module): common 모듈 분리
각 도메인들에서 공통으로 사용하는 도메인을 모아둔 모듈
