+++ kitchenpos-common/src/main/java/kitchenpos/common/domain/Empty.java
