+++ kitchenpos-common/src/main/java/kitchenpos/common/domain/NumberOfGuests.java
