+++ kitchenpos-common/src/main/java/kitchenpos/common/domain/Price.java
