+++ kitchenpos-common/src/main/java/kitchenpos/common/domain/Quantity.java
