+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/EmptyTest.java
