+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/NameTest.java
