+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/NumberOfGuestsTest.java
