+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/PriceTest.java
