+++ kitchenpos-common/src/test/java/kitchenpos/common/domain/QuantityTest.java
