+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -2,7 +2,6 @@ package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.product.domain.Product;
 import kitchenpos.common.domain.Quantity;
 
 public class MenuProductResponse {
