+++ src/test/java/kitchenpos/order/acceptance/OrderAcceptanceTest.java
@@ -1,19 +1,19 @@
 package kitchenpos.order.acceptance;
 
 import static java.util.Collections.singletonList;
-import static kitchenpos.menugroup.acceptance.MenuGroupRestAssured.메뉴_그룹_등록되어_있음;
 import static kitchenpos.menu.acceptance.MenuRestAssured.메뉴_등록되어_있음;
-import static kitchenpos.menugroup.domain.MenuGroupTestFixture.generateMenuGroupRequest;
 import static kitchenpos.menu.domain.MenuProductTestFixture.generateMenuProductRequest;
 import static kitchenpos.menu.domain.MenuTestFixture.generateMenuRequest;
+import static kitchenpos.menugroup.acceptance.MenuGroupRestAssured.메뉴_그룹_등록되어_있음;
+import static kitchenpos.menugroup.domain.MenuGroupTestFixture.generateMenuGroupRequest;
 import static kitchenpos.order.acceptance.OrderRestAssured.주문_등록되어_있음;
 import static kitchenpos.order.acceptance.OrderRestAssured.주문_목록_조회_요청;
 import static kitchenpos.order.acceptance.OrderRestAssured.주문_상태_변경_요청;
 import static kitchenpos.order.acceptance.OrderRestAssured.주문_생성_요청;
-import static kitchenpos.ordertable.acceptance.TableRestAssured.주문_테이블_등록되어_있음;
 import static kitchenpos.order.domain.OrderLineItemTestFixture.generateOrderLineItemRequest;
-import static kitchenpos.ordertable.domain.OrderTableTestFixture.generateOrderTableRequest;
 import static kitchenpos.order.domain.OrderTestFixture.generateOrderRequest;
+import static kitchenpos.ordertable.acceptance.TableRestAssured.주문_테이블_등록되어_있음;
+import static kitchenpos.ordertable.domain.OrderTableTestFixture.generateOrderTableRequest;
 import static kitchenpos.product.acceptance.ProductRestAssured.상품_등록되어_있음;
 import static kitchenpos.product.domain.ProductTestFixture.generateProductRequest;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -26,10 +26,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.common.acceptance.AcceptanceTest;
-import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.AcceptanceTest;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuResponse;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
